move command to debug menu and change wording to “Open Preferences File”
