+++ src/help/HelpCommandHandlers.js
@@ -40,9 +40,7 @@ define(function (require, exports, module) {
         NativeApp               = require("utils/NativeApp"),
         StringUtils             = require("utils/StringUtils"),
         AboutDialogTemplate     = require("text!htmlContent/about-dialog.html"),
-        ContributorsTemplate    = require("text!htmlContent/contributors-list.html"),
-        PreferencesManager      = require("preferences/PreferencesManager"),
-        FileSystem              = require("filesystem/FileSystem");
+        ContributorsTemplate    = require("text!htmlContent/contributors-list.html");
     
     var buildInfo;
     
@@ -65,22 +63,6 @@ define(function (require, exports, module) {
             FileUtils.convertToNativePath(decodeURI(window.location.href)),
             function (err) {} /* Ignore errors */
         );
-    }
-    
-    function _handleOpenSettings() {
-        var fullPath = PreferencesManager.getUserPrefFile(),
-            file = FileSystem.getFileForPath(fullPath);
-        file.exists(function (err, doesExist) {
-            if (doesExist) {
-                CommandManager.execute(Commands.FILE_OPEN, { fullPath: fullPath });
-            } else {
-                FileUtils.writeText(file, "", true)
-                    .done(function () {
-                        CommandManager.execute(Commands.FILE_OPEN, { fullPath: fullPath });
-                    });
-            }
-        });
-        
     }
 
     function _handleAboutDialog() {
@@ -149,5 +131,4 @@ define(function (require, exports, module) {
     CommandManager.register(Strings.CMD_SHOW_EXTENSIONS_FOLDER, Commands.HELP_SHOW_EXT_FOLDER,      _handleShowExtensionsFolder);
     CommandManager.register(Strings.CMD_TWITTER,                Commands.HELP_TWITTER,              _handleLinkMenuItem(brackets.config.twitter_url));
     CommandManager.register(Strings.CMD_ABOUT,                  Commands.HELP_ABOUT,                _handleAboutDialog);
-    CommandManager.register(Strings.CMD_OPEN_SETTINGS,          Commands.HELP_OPEN_SETTINGS,             _handleOpenSettings);
 });
